Re-factored into separate class that isolates the logic for fixing ids Creating additional test steps in unit tests Fix *all* invalid ids in references, as well as actual resources
